SPY-48: Incr/Decr param "by" should be able to take a long
Reviewed-on: http://review.couchbase.org/9389
Tested-by: Michael Wiederhold <mike@couchbase.com>

+++ src/main/java/net/spy/memcached/OperationFactory.java
@@ -146,7 +146,7 @@ public interface OperationFactory {
    * @param cb the status callback
    * @return the new mutator operation
    */
-  MutatorOperation mutate(Mutator m, String key, int by, long def, int exp,
+  MutatorOperation mutate(Mutator m, String key, long by, long def, int exp,
       OperationCallback cb);
 
   /**
