+++ src/main/java/net/spy/memcached/protocol/binary/BinaryOperationFactory.java
@@ -94,7 +94,7 @@ public class BinaryOperationFactory extends BaseOperationFactory {
     return new GetsOperationImpl(key, cb);
   }
 
-  public MutatorOperation mutate(Mutator m, String key, int by, long def,
+  public MutatorOperation mutate(Mutator m, String key, long by, long def,
       int exp, OperationCallback cb) {
     return new MutatorOperationImpl(m, key, by, def, exp, cb);
   }
